+++ backend/main.py
@@ -6,7 +6,7 @@ Look in api.routes for the actual api code
 """
 
 from multiprocessing import Process, Queue
-from microlab import startMicrolabProcess
+from microlab.core import startMicrolabProcess
 from api.core import runFlask
 import config
 import multiprocessing_logging
